# lib/api-zod/src/generated/types/downloadInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pawchive API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadInput {
  post_id: number;
  /** @nullable */
  media_item_id?: number | null;
}