# lib/api-zod/src/generated/types/downloadStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pawchive API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadStats {
  total: number;
  queued: number;
  running: number;
  done: number;
  error: number;
}