# lib/api-zod/src/generated/types/downloadTask.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pawchive API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadTask {
  id: number;
  post_id: number;
  /** @nullable */
  post_title?: string | null;
  /** @nullable */
  author?: string | null;
  /** @nullable */
  media_item_id?: number | null;
  /** @nullable */
  filename?: string | null;
  /** queued | running | done | error */
  status: string;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  error_message?: string | null;
  /** @nullable */
  local_path?: string | null;
  created_at: string;
  /** @nullable */
  completed_at?: string | null;
}