# lib/api-zod/src/generated/types/libraryAsset.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pawchive API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LibraryAsset {
  id: number;
  /** @nullable */
  post_id?: number | null;
  /** @nullable */
  post_title?: string | null;
  /** @nullable */
  author?: string | null;
  filename: string;
  media_type: string;
  /** @nullable */
  local_path?: string | null;
  /** @nullable */
  file_size?: number | null;
  /** available | missing */
  status: string;
  /** @nullable */
  thumbnail_url?: string | null;
  created_at: string;
}