# lib/api-zod/src/generated/types/scanLogEntry.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pawchive API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScanLogEntry {
  timestamp: string;
  /** info | warn | error */
  level: string;
  message: string;
}