# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pawchive API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanLogEntry } from './scanLogEntry';

export interface ScanResult {
  discovered: number;
  source_name: string;
  new_posts?: number;
  log?: ScanLogEntry[];
}