# lib/api-zod/src/generated/types/settingsUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pawchive API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SettingsUpdate {
  download_path?: string;
  library_path?: string;
  max_concurrent_downloads?: number;
  auto_scan?: boolean;
  /** @nullable */
  auto_scan_interval?: number | null;
  /** @nullable */
  patreon_access_token?: string | null;
}